# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FlowPilot API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects": {
      "get": {
        "tags": [
          "Workspace"
        ],
        "summary": "Projects",
        "operationId": "projects_api_v1_projects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Workspace"
        ],
        "summary": "Add Project",
        "operationId": "add_project_api_v1_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows": {
      "get": {
        "tags": [
          "Workflows"
        ],
        "summary": "Workflows",
        "operationId": "workflows_api_v1_workflows_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Workflows"
        ],
        "summary": "Add Workflow",
        "operationId": "add_workflow_api_v1_workflows_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/workflows/{workflow_id}/publish": {
      "post": {
        "tags": [
          "Workflows"
        ],
        "summary": "Publish Workflow",
        "operationId": "publish_workflow_api_v1_workflows__workflow_id__publish_post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Publish"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/workflows/{workflow_id}/runs": {
      "post": {
        "tags": [
          "Executions"
        ],
        "summary": "Start Run",
        "operationId": "start_run_api_v1_workflows__workflow_id__runs_post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 120,
                  "pattern": "^[A-Za-z0-9_.:-]+$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/runs/{run_id}": {
      "get": {
        "tags": [
          "Executions"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AiNode": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9_]{0,47}$",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "maxLength": 80,
            "title": "Label",
            "default": ""
          },
          "depends_on": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 32,
            "title": "Depends On"
          },
          "join": {
            "type": "string",
            "enum": [
              "all",
              "any"
            ],
            "title": "Join",
            "default": "all"
          },
          "when": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Predicate"
              },
              {
                "type": "null"
              }
            ]
          },
          "retry": {
            "$ref": "#/components/schemas/RetryPolicy"
          },
          "timeout_seconds": {
            "type": "integer",
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Timeout Seconds",
            "default": 15
          },
          "type": {
            "type": "string",
            "const": "ai",
            "title": "Type"
          },
          "provider": {
            "type": "string",
            "enum": [
              "mock",
              "openai",
              "anthropic",
              "gemini"
            ],
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "maxLength": 100,
            "pattern": "^[A-Za-z0-9._:/-]+$",
            "title": "Model",
            "default": "fixture"
          },
          "prompt": {
            "title": "Prompt"
          },
          "credential_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credential Id"
          },
          "response_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Schema"
          },
          "mock_response": {
            "title": "Mock Response"
          },
          "max_output_tokens": {
            "type": "integer",
            "maximum": 2048.0,
            "minimum": 16.0,
            "title": "Max Output Tokens",
            "default": 512
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "provider",
          "prompt"
        ],
        "title": "AiNode"
      },
...
      "ApprovalNode": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9_]{0,47}$",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "maxLength": 80,
            "title": "Label",
            "default": ""
          },
          "depends_on": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 32,
            "title": "Depends On"
          },
          "join": {
            "type": "string",
            "enum": [
              "all",
              "any"
            ],
            "title": "Join",
            "default": "all"
          },
          "when": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Predicate"
              },
              {
                "type": "null"
              }
            ]
          },
          "retry": {
            "$ref": "#/components/schemas/RetryPolicy"
          },
          "timeout_seconds": {
            "type": "integer",
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Timeout Seconds",
            "default": 15
          },
          "type": {
            "type": "string",
            "const": "approval",
            "title": "Type"
          },
          "message": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Message"
          },
          "expires_after": {
            "type": "integer",
            "maximum": 604800.0,
            "minimum": 60.0,
            "title": "Expires After",
            "default": 86400
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "message"
        ],
        "title": "ApprovalNode"
      },
      "ConditionNode": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9_]{0,47}$",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "maxLength": 80,
            "title": "Label",
            "default": ""
          },
          "depends_on": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 32,
            "title": "Depends On"
          },
          "join": {
            "type": "string",
            "enum": [
              "all",
              "any"
            ],
            "title": "Join",
            "default": "all"
          },
          "when": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Predicate"
              },
              {
                "type": "null"
              }
            ]
          },
          "retry": {
            "$ref": "#/components/schemas/RetryPolicy"
          },
          "timeout_seconds": {
            "type": "integer",
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Timeout Seconds",
            "default": 15
          },
          "type": {
            "type": "string",
            "const": "condition",
            "title": "Type"
          },
          "predicate": {
            "$ref": "#/components/schemas/Predicate"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "predicate"
        ],
        "title": "ConditionNode"
      },
...
      "Definition": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version",
            "default": 1
          },
          "input_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Schema"
          },
          "nodes": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/SetNode"
                },
                {
                  "$ref": "#/components/schemas/ConditionNode"
                },
                {
                  "$ref": "#/components/schemas/DelayNode"
                },
                {
                  "$ref": "#/components/schemas/ApprovalNode"
                },
                {
                  "$ref": "#/components/schemas/HttpNode"
                },
                {
                  "$ref": "#/components/schemas/AiNode"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "ai": "#/components/schemas/AiNode",
                  "approval": "#/components/schemas/ApprovalNode",
                  "condition": "#/components/schemas/ConditionNode",
                  "delay": "#/components/schemas/DelayNode",
                  "http": "#/components/schemas/HttpNode",
                  "set": "#/components/schemas/SetNode"
                }
              }
            },
            "type": "array",
            "maxItems": 32,
            "minItems": 1,
            "title": "Nodes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "nodes"
        ],
        "title": "Definition"
      },
      "DelayNode": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9_]{0,47}$",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "maxLength": 80,
            "title": "Label",
            "default": ""
          },
          "depends_on": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 32,
            "title": "Depends On"
          },
          "join": {
            "type": "string",
            "enum": [
              "all",
              "any"
            ],
            "title": "Join",
            "default": "all"
          },
          "when": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Predicate"
              },
              {
                "type": "null"
              }
            ]
          },
          "retry": {
            "$ref": "#/components/schemas/RetryPolicy"
          },
          "timeout_seconds": {
            "type": "integer",
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Timeout Seconds",
            "default": 15
          },
          "type": {
            "type": "string",
            "const": "delay",
            "title": "Type"
          },
          "seconds": {
            "type": "integer",
            "maximum": 604800.0,
            "minimum": 1.0,
            "title": "Seconds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "seconds"
        ],
        "title": "DelayNode"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HttpNode": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9_]{0,47}$",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "maxLength": 80,
            "title": "Label",
            "default": ""
          },
          "depends_on": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 32,
            "title": "Depends On"
          },
          "join": {
            "type": "string",
            "enum": [
              "all",
              "any"
            ],
            "title": "Join",
            "default": "all"
          },
          "when": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Predicate"
              },
              {
                "type": "null"
              }
            ]
          },
          "retry": {
            "$ref": "#/components/schemas/RetryPolicy"
          },
          "timeout_seconds": {
            "type": "integer",
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Timeout Seconds",
            "default": 15
          },
          "type": {
            "type": "string",
            "const": "http",
            "title": "Type"
          },
          "url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 9,
            "title": "Url"
          },
          "method": {
            "type": "string",
            "enum": [
              "GET",
              "POST",
              "PUT",
              "PATCH",
              "DELETE"
            ],
            "title": "Method",
            "default": "GET"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Headers"
          },
          "body": {
            "title": "Body"
          },
          "credential_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credential Id"
          },
          "receiver_supports_idempotency": {
            "type": "boolean",
            "title": "Receiver Supports Idempotency",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "url"
        ],
        "title": "HttpNode"
      },
      "Login": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "minLength": 3,
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "format": "password",
            "title": "Password",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "Login"
      },
...
      "Predicate": {
        "properties": {
          "left": {
            "title": "Left"
          },
          "op": {
            "type": "string",
            "enum": [
              "eq",
              "ne",
              "gt",
              "gte",
              "lt",
              "lte",
              "contains",
              "exists"
            ],
            "title": "Op",
            "default": "eq"
          },
          "right": {
            "title": "Right"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "left"
        ],
        "title": "Predicate"
      },
      "ProjectCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProjectCreate"
      },
      "Publish": {
        "properties": {
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision"
        ],
        "title": "Publish"
      },
...
      "RetryPolicy": {
        "properties": {
          "max_attempts": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Max Attempts",
            "default": 3
          },
          "initial_seconds": {
            "type": "number",
            "maximum": 30.0,
            "minimum": 0.1,
            "title": "Initial Seconds",
            "default": 1
          },
          "max_seconds": {
            "type": "number",
            "maximum": 300.0,
            "minimum": 1.0,
            "title": "Max Seconds",
            "default": 60
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RetryPolicy"
      },
      "RunCreate": {
        "properties": {
          "input": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RunCreate"
      },
      "SetNode": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9_]{0,47}$",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "maxLength": 80,
            "title": "Label",
            "default": ""
          },
          "depends_on": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 32,
            "title": "Depends On"
          },
          "join": {
            "type": "string",
            "enum": [
              "all",
              "any"
            ],
            "title": "Join",
            "default": "all"
          },
          "when": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Predicate"
              },
              {
                "type": "null"
              }
            ]
          },
          "retry": {
            "$ref": "#/components/schemas/RetryPolicy"
          },
          "timeout_seconds": {
            "type": "integer",
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Timeout Seconds",
            "default": 15
          },
          "type": {
            "type": "string",
            "const": "set",
            "title": "Type"
          },
          "values": {
            "additionalProperties": true,
            "type": "object",
            "title": "Values"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "values"
        ],
        "title": "SetNode"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkflowCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "title": "Description",
            "default": ""
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "environment": {
            "type": "string",
            "enum": [
              "development",
              "staging",
              "production"
            ],
            "title": "Environment",
            "default": "development"
          },
          "definition": {
            "$ref": "#/components/schemas/Definition"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "project_id",
          "definition"
        ],
        "title": "WorkflowCreate"
      }
    }
  }
}